+++ src/main/java/gs/mclo/api/internal/filter/Filter.java
@@ -1,0 +1,9 @@
+package gs.mclo.api.internal.filter;
+
+import org.jetbrains.annotations.ApiStatus;
+
+@ApiStatus.NonExtendable
+public interface Filter {
+    String getType();
+    String apply(String input);
+}