+++ src/main/java/gs/mclo/api/internal/filter/TrimFilter.java
@@ -1,0 +1,13 @@
+package gs.mclo.api.internal.filter;
+
+public final class TrimFilter implements Filter {
+    @Override
+    public String getType() {
+        return "trim";
+    }
+
+    @Override
+    public String apply(String input) {
+        return input.trim();
+    }
+}