+++ src/test/java/gs/mclo/api/LimitedReaderTest.java
@@ -1,7 +1,6 @@
 package gs.mclo.api;
 
 import gs.mclo.api.internal.LimitedReader;
-import gs.mclo.api.response.Limits;
 import org.junit.jupiter.api.Test;
 
 import java.io.StringReader;
@@ -13,7 +12,7 @@ public class LimitedReaderTest {
     void testReadSimpleString() {
         assertDoesNotThrow(() -> {
             String input = "Hello, world!";
-            try (LimitedReader reader = new LimitedReader(new StringReader(input), new Limits(1, 100, 100))) {
+            try (LimitedReader reader = new LimitedReader(new StringReader(input), 100, 100, true)) {
                 char[] buffer = new char[100];
 
                 int offset = 0;
@@ -43,7 +42,7 @@ public class LimitedReaderTest {
     void testReadByteLimit() {
         assertDoesNotThrow(() -> {
             String input = "Hello, world!";
-            try (LimitedReader reader = new LimitedReader(new StringReader(input), new Limits(1, 10, 100))) {
+            try (LimitedReader reader = new LimitedReader(new StringReader(input), 10, 100, true)) {
                 char[] buffer = new char[100];
 
                 int offset = 0;
@@ -73,7 +72,7 @@ public class LimitedReaderTest {
     void testReadByteLimitUnicode() {
         assertDoesNotThrow(() -> {
             String input = "Hello, world 🌍!";
-            try (LimitedReader reader = new LimitedReader(new StringReader(input), new Limits(1, 17, 100))) {
+            try (LimitedReader reader = new LimitedReader(new StringReader(input), 17, 100, true)) {
                 char[] buffer = new char[100];
 
                 int offset = 0;
@@ -103,7 +102,7 @@ public class LimitedReaderTest {
     void testReadByteLimitUnicodeCutoff() {
         assertDoesNotThrow(() -> {
             String input = "Hello, world 🌍";
-            try (LimitedReader reader = new LimitedReader(new StringReader(input), new Limits(1, 14, 100))) {
+            try (LimitedReader reader = new LimitedReader(new StringReader(input), 14, 100, true)) {
                 char[] buffer = new char[100];
 
                 int offset = 0;
@@ -133,7 +132,7 @@ public class LimitedReaderTest {
     void testReadLineLimit() {
         assertDoesNotThrow(() -> {
             String input = "H\ne\nl\nl\no, world!";
-            try (LimitedReader reader = new LimitedReader(new StringReader(input), new Limits(1, 100, 3))) {
+            try (LimitedReader reader = new LimitedReader(new StringReader(input), 100, 3, true)) {
                 char[] buffer = new char[100];
 
                 int offset = 0;