+++ src/test/java/gs/mclo/api/LogTest.java
@@ -4,6 +4,8 @@ import gs.mclo.api.data.Metadata;
 import gs.mclo.api.reader.LogReader;
 import gs.mclo.api.reader.StringLogReader;
 import gs.mclo.api.response.Limits;
+import gs.mclo.api.internal.filter.Filter;
+import gs.mclo.api.internal.filter.FilterList;
 import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.Test;
 
@@ -12,50 +14,52 @@ import java.util.Set;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-public class LogTest {
+public class LogTest extends ApiTest {
     private static final String REMOVED_IPv6 = "****:****:****:****:****:****:****:****";
     private static final String REMOVED_IPv4 = "**.**.**.**";
 
+    private final FilterList filters = new FilterList(new Filter[]{});
+
     @Test
     void removeIPv4() {
-        assertEquals(REMOVED_IPv4, filter("123.45.67.89"));
+        assertFilter(REMOVED_IPv4, "123.45.67.89");
     }
 
     @Test
     void removeIPv4s() {
-        assertEquals(REMOVED_IPv4 + " " + REMOVED_IPv4, filter("123.45.67.89 189.123.42.34"));
+        assertFilter(REMOVED_IPv4 + " " + REMOVED_IPv4, "123.45.67.89 189.123.42.34");
     }
 
     @Test
     void similarToIPv4() {
-        assertEquals("1000.45.67.89 " + REMOVED_IPv4, filter("1000.45.67.89 189.123.42.34"));
+        assertFilter("1000.45.67.89 " + REMOVED_IPv4, "1000.45.67.89 189.123.42.34");
     }
 
     @Test
     void whitelistedIpv4s() {
         String content = "127.0.0.1 127.75.75.18 0.0.0.0 1.1.1.1 1.0.0.1 8.8.8.8 8.8.8.4";
-        assertEquals(content, filter(content));
+        assertFilter(content, content);
     }
 
     @Test
     void removeIPv6() {
-        assertEquals(REMOVED_IPv6, filter("9557:c600:a213:4835:fdaa:9d04:e354:ed9e"));
+        assertFilter(REMOVED_IPv6, "9557:c600:a213:4835:fdaa:9d04:e354:ed9e");
     }
 
     @Test
     void removeIPv6s() {
-        assertEquals(REMOVED_IPv6 + " " + REMOVED_IPv6, filter("010b:0611:2138:c376:6c8f:a46e:af48:7014 53a6:8214:7341:e156:0c3c:ffcc:3474:e207"));
+        assertFilter(REMOVED_IPv6 + " " + REMOVED_IPv6, "010b:0611:2138:c376:6c8f:a46e:af48:7014 53a6:8214:7341:e156:0c3c:ffcc:3474:e207");
     }
 
     @Test
     void similarToIPv6() {
-        assertEquals("10000:8214:7341:e156:0c3c:ffcc:3474:e207", filter("10000:8214:7341:e156:0c3c:ffcc:3474:e207"));
+        assertFilter("10000:8214:7341:e156:0c3c:ffcc:3474:e207", "10000:8214:7341:e156:0c3c:ffcc:3474:e207");
     }
 
     @Test
     void whitelistedIpv6s() {
         String content = "::1 0:0:0:0:0:0:0:0:1 ::0 0:0:0:0:0:0:0:0:0";
-        assertEquals(content, filter(content));
+        assertFilter(content, content);
     }
 
     @Test
@@ -77,8 +81,8 @@ public class LogTest {
         var log = new Log(testReader);
 
         try {
-            String firstCall = log.getContent(Limits.DEFAULT);
-            String secondCall = log.getContent(Limits.DEFAULT);
+            String firstCall = log.getContent(filters);
+            String secondCall = log.getContent(filters);
             assertEquals("Test content", firstCall);
             assertEquals("Test content", secondCall);
         } catch (IOException e) {
@@ -118,9 +122,22 @@ public class LogTest {
         assertNull(log.getSource());
     }
 
-    protected String filter(String content) {
+    protected void assertFilter(String expected, String content) {
+        assertEquals(expected, legacyFilter(content));
+        assertEquals(expected, filter(content));
+    }
+
+    protected String legacyFilter(String content) {
         try {
             return new Log(new StringLogReader(content)).getContent(Limits.DEFAULT);
+        } catch (IOException e) {
+            throw new UncheckedIOException(e);
+        }
+    }
+
+    protected String filter(String content) {
+        try {
+            return new Log(new StringLogReader(content)).getContent(client.getFilters().join());
         } catch (IOException e) {
             throw new UncheckedIOException(e);
         }